refactor(ui): sse updates
