+++ services/dashboard-ui/client/components/builds/BuildTimeline/BuildTimelineContainer.tsx
@@ -1,7 +1,8 @@
 import { useSearchParams } from 'react-router'
 import { useApp } from '@/hooks/use-app'
 import { useOrg } from '@/hooks/use-org'
-import { useSSETimelineQuery } from '@/hooks/use-sse-timeline-query'
+import { useSSETimelineQuery } from '@/lib/sse/use-sse-timeline-query'
+import { useRefreshErrorToast } from '@/hooks/use-refresh-error-toast'
 import { getComponentBuilds } from '@/lib'
 import { BuildTimeline } from './BuildTimeline'
 
@@ -25,6 +26,8 @@ export const BuildTimelineContainer = ({
   const [searchParams] = useSearchParams()
   const offset = Number(searchParams.get('offset') ?? 0)
 
+  const onRefreshError = useRefreshErrorToast()
+
   const { data: result } = useSSETimelineQuery({
     sseUrl:
       org?.id && componentId
@@ -42,6 +45,7 @@ export const BuildTimelineContainer = ({
     shouldPoll,
     pollInterval,
     eventName: 'builds',
+    onError: onRefreshError,
   })
 
   const builds = result?.data ?? []
