+++ services/dashboard-ui/client/components/deploys/DeployTimeline/DeployTimelineContainer.tsx
@@ -1,7 +1,8 @@
 import { useSearchParams } from 'react-router'
 import { useInstall } from '@/hooks/use-install'
 import { useOrg } from '@/hooks/use-org'
-import { useSSETimelineQuery } from '@/hooks/use-sse-timeline-query'
+import { useSSETimelineQuery } from '@/lib/sse/use-sse-timeline-query'
+import { useRefreshErrorToast } from '@/hooks/use-refresh-error-toast'
 import { getComponentDeploys } from '@/lib'
 import { DeployTimeline } from './DeployTimeline'
 
@@ -25,6 +26,8 @@ export const DeployTimelineContainer = ({
   const [searchParams] = useSearchParams()
   const offset = Number(searchParams.get('offset') ?? 0)
 
+  const onRefreshError = useRefreshErrorToast()
+
   const { data: result, isLoading, error } = useSSETimelineQuery({
     sseUrl:
       org?.id && install?.id && componentId
@@ -43,6 +46,7 @@ export const DeployTimelineContainer = ({
     shouldPoll,
     pollInterval,
     eventName: 'deploys',
+    onError: onRefreshError,
   })
 
   const deploys = result?.data ?? []
