+++ services/dashboard-ui/client/hooks/use-refresh-error-toast.tsx
@@ -1,19 +1,19 @@
-import { useEffect } from 'react'
+import { useCallback } from 'react'
 import { useToast } from '@/hooks/use-toast'
 import { Text } from '@/components/common/Text'
 import { Toast } from '@/components/surfaces/Toast'
-import type { TAPIError } from '@/types'
 
-export function useRefreshErrorToast(error: unknown, data: unknown) {
+export function useRefreshErrorToast() {
   const { addToast } = useToast()
 
-  useEffect(() => {
-    if (error && data) {
+  return useCallback(
+    (message: string) => {
       addToast(
         <Toast heading="Refresh failed" theme="warn">
-          <Text>{(error as TAPIError)?.error ?? 'Connection issue'}</Text>
+          <Text>{message}</Text>
         </Toast>
       )
-    }
-  }, [error])
+    },
+    [addToast]
+  )
 }
